Add query-retrieve script based on dcm4che tools v3
This script provides a way to list the entire DB of a PACS server in
json format.
It can also download all studies by patient.
It allows to approach the problem from a higher level perspective to guide
the structure of the go-dicom development to eventually replace dcm4che
with it.